Overlap between Two Empirical Density Estimates

This function generates kernel density estimates for two datasets on a common grid to calculate the area of overlap between the two estimates. See http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities, this function is derived from code posted in the answer by user mmk.

pairwise_overlap(a, b, density_args = list())

Details

FIXME new documentation could be added here, if we want to still export this function.